+++ ViewModels/ItemSalesData.cs
@@ -5,7 +5,7 @@
         public int ItemId { get; set; }
         public string ItemName { get; set; }
         public string ItemCode { get; set; }
-        public int QuantitySold { get; set; }
+        public decimal QuantitySold { get; set; }
         public decimal TotalSales { get; set; }
         public decimal TotalCost { get; set; }
         public decimal Profit => TotalSales - TotalCost;
